Handle Excluding options From distinct + count
If you pass callback as options, then the underlying connector will detect it as the callback, and it'll get excuted with the results of query, and our handler will never get executed. That's pretty lame and hard to diagnose, so let's detect the scenario and handle it properly. Pow pow, take that, bug.